重构商品详情导航逻辑

+++ core/model/src/main/java/com/joker/coolmall/core/model/Goods.kt
@@ -16,7 +16,7 @@ data class Goods(
     /**
      * 类型ID
      */
-    val typeId: Int = 0,
+    val typeId: Long = 0,
 
     /**
      * 标题
